error bars are SME. Feat: show distributions
The user can toggle the show_distributions attribute. This will plot the moments for each cluster at each cluster size.
This is useful if you want to better understand the distribution of moments instead of just showing the scaling of the mean.
This part of the code should still be improved and is a bit messy at the moment.